==> real_data/data_preprocess/00_2_extract_flowline.R <==
## Extract data along flowline

library(dplyr)
library(sf)
library(ggplot2)
library(qs)

extract_flowline <- F
regrid_flowline <- F

## Read (averaged) velocity data
setwd("~/SSA_model/CNN/real_data/")
data_dir <- "data/"

basin_data <- read_sf(paste0(data_dir, "/boundaries/Basins/Basins_Antarctica_v02.shp"))
thwaites_bound <- basin_data %>% filter(NAME == "Thwaites")
thwaites_points <- as.data.frame(st_coordinates(st_geometry(thwaites_bound)))
  
## Velocity data on Thwaites glacier
  vel_thwaites <- qread(paste0(data_dir, "velocity/vel_thwaites.qs"))
  vel_thwaites <- vel_thwaites %>% filter(!is.na(v)) # remove NA velocities
  
  ## Choose a point in the middle of the glacier
  half_range_x <- min(thwaites_points$X) + 200e3 #400e3 
  half_range_y <- min(thwaites_points$Y) + 450e3 #500e3 

  x_unique <- unique(vel_thwaites$x) # x-coordinates
  delta <- x_unique[2] - x_unique[1] # grid size

  midpts <- vel_thwaites %>%
    filter(x >= (half_range_x - delta) & x <= half_range_x + delta & 
          y >= (half_range_y - delta) & y <= (half_range_y + delta))

  # head(midpts)
  chosen_pt <- midpts[1, ]

  ## Plot data around the chosen point
  margin <- 100e3
  near_pts <- vel_thwaites #%>% filter(x >= (chosen_pt$x - margin) & x <= (chosen_pt$x + margin) &
    # y >= (chosen_pt$y - margin) & y <= (chosen_pt$y + margin))

  png(paste0("./plots/flowline_start_point.png"), width = 800, height = 800)
  ggplot() +
    # geom_point(data = near_pts, aes(x = x, y = y, colour = v)) +
    # scale_colour_distiller(
    #   palette = "Reds", direction = 1,
    #   name = "Velocity (m/a)"
    # ) +
    geom_sf(data = thwaites_bound, color = "black", fill = NA) +
    geom_point(data = data.frame(chosen_pt), aes(x = x, y = y), color = "black", size = 5) +
    theme_bw()
  dev.off()


if (extract_flowline) {

  stepsize <- 1

  # Extract data along flowline
  print("Extracting flowline...")
  pos <- as.numeric(chosen_pt[c("x", "y")])

  positions <- list()
  positions[[1]] <- pos
  # velocities <- list()
  # velocities[[1]] <- as.numeric(chosen_pt[c("v", "vx", "vy")])

  while (pos[1] > (min(vel_thwaites$x) + delta) & pos[2] > (min(vel_thwaites$y) + delta)) {
    # for (i in 1:10) {

    ## need to find 4 nearest grid points around the current position
    near_pts <- vel_thwaites %>% filter(
      x >= (pos[1] - delta) & x <= (pos[1] + delta),
      y >= (pos[2] - delta) & y <= (pos[2] + delta)
    )


    near_v <- near_pts %>%
      mutate(dist = sqrt((x - pos[1])^2 + (y - pos[2])^2)) %>%
      filter(dist > 0) %>%
      arrange(dist) %>%
      slice_min(dist, n = 4) %>%
      select(v, vx, vy) %>%
      summarise(v_nearest = v[1], v_avg = mean(v), vx_avg = mean(vx), vy_avg = mean(vy)) # %>% # take average vx and vx as vx and vy at current position
    # as.numeric()

    vx_vy <- as.numeric(near_v[c("vx_avg", "vy_avg")])

    if (is.na(vx_vy[1])) {
      break
    }

    cat("vx_vy = ", vx_vy, "\n")

    pos <- pos + stepsize * vx_vy
    positions <- c(positions, list(pos))
    print(pos)

    # velocities <- c(velocities, list(near_v))
  }

  # saveRDS(positions, paste0(data_dir, "/flowline_positions.rds"))
  qsave(positions, paste0(data_dir, "/flowline_positions.qs"))
} else {
  # positions <- readRDS(paste0(data_dir, "/flowline_positions.rds"))
  positions <- qread(paste0(data_dir, "/flowline_positions.qs"))
}

# positions
flowline_x <- sapply(positions, function(x) x[1])
flowline_y <- sapply(positions, function(x) x[2])

# gl_thwaites <- readRDS(paste0(data_dir, "/gl_thwaites.rds"))
# gl_thwaites <- qread(paste0(data_dir, "/gl_thwaites.qs"))

plot_flowline <- ggplot(vel_thwaites) +
    # geom_point(aes(x = x, y = y, colour = log10(v))) +
    geom_point(aes(x = x, y = y, colour = v)) +
    scale_colour_distiller(
      palette = "BuPu", direction = 1,
      limits = c(1, 10000),
      name = "Velocity (m/yr)"
    ) +
    # geom_point(data = gl_thwaites, aes(x = X, y = Y), color = "black", size = 0.2) +
    geom_sf(data = thwaites_bound, color = "black", fill = NA) +
    geom_point(data = data.frame(chosen_pt), aes(x = x, y = y), color = "black", size = 3) +
    geom_line(data = data.frame(x = flowline_x, y = flowline_y),
                aes(x = x, y = y), color = "red", linewidth = 1) +
    theme_bw() +
    theme(text = element_text(size = 16), 
          axis.title = element_blank()) 

print("Saving flowline plot...")
png(paste0("./plots/flowline2.png"), width = 800, height = 800, res = 100)
print(plot_flowline)
# plot(flowline_x, flowline_y, type = "l", col = "blue", xlab = "x", ylab = "y")
dev.off()

browser()

## Re-grid
if (regrid_flowline) {
  print("Re-gridding flowline...")
  sq_x_diff <- (flowline_x[2:length(flowline_x)] - flowline_x[1:(length(flowline_x) - 1)])^2
  sq_y_diff <- (flowline_y[2:length(flowline_y)] - flowline_y[1:(length(flowline_y) - 1)])^2
  dist <- sqrt(sq_x_diff + sq_y_diff)
  flowline_length <- sum(dist)

  J <- 2001 #2001
  grid_size <- flowline_length/J

  cumul_dist <- c(0, cumsum(dist))
  # cumul_dist <- lapply(1:5, function(i) cumsum(dist[1:i]))

  which.closest <- function(x, vec) {
    sub_vec <- vec[vec <= x]
    which.min(x - sub_vec)
  }

  ini_pt <- c(flowline_x[1], flowline_y[1])

  x_new <- c()
  y_new <- c()
  for (j in 1:J) {
    closest <- which.closest(grid_size * j, cumul_dist)
    leftover <- grid_size * j - cumul_dist[closest]

    l <- leftover / dist[closest]
    m <- 1 - l

    x_new[j] <- (m * flowline_x[closest] + l * flowline_x[closest + 1]) 
    y_new[j] <- (m * flowline_y[closest] + l * flowline_y[closest + 1])

  }

  x_new <- c(flowline_x[1], x_new) ## add initial point
  y_new <- c(flowline_y[1], y_new)

  ## Plot re-gridded flowline
  plot_range <- (length(flowline_x) - 10):length(flowline_x)
  plot(flowline_x[plot_range], flowline_y[plot_range], type = "o", xlab = "x", ylab = "y")
  points(x_new, y_new, col = "red", pch = 16)

  flowline_regrid <- data.frame(x = x_new, y = y_new)
  flowline_regrid <- na.omit(flowline_regrid)

  ## Truncate flowline to 2001 points
  # flowline_regrid <- flowline_regrid[1:2001, ]

  print("Saving re-gridded flowline...")
  # saveRDS(flowline_regrid, paste0(data_dir, "/flowline_regrid.rds"))
  qsave(flowline_regrid, paste0(data_dir, "/flowline_regrid.qs"))
} else {
  #  flowline_regrid <- readRDS(paste0(data_dir, "/flowline_regrid.rds"))
    flowline_regrid <- qread(paste0(data_dir, "/flowline_regrid.qs"))
}

plot_flowline_rg <- ggplot(vel_thwaites) +
    geom_point(aes(x = x, y = y, colour = log10(v))) +
    scale_colour_distiller(
      palette = "BuPu", direction = 1,
      # limits = log10(c(1, 2000)),
      name = "Log-velocity (m/a)"
    ) +
    geom_point(data = gl_thwaites, aes(x = X, y = Y), color = "salmon", size = 0.5) +
    geom_sf(data = thwaites_bound, color = "black", fill = NA) +
    geom_point(data = data.frame(chosen_pt), aes(x = x, y = y), color = "black", size = 5) +
    geom_point(data = data.frame(x = flowline_x, y = flowline_y),
                aes(x = x, y = y), color = "black") +
    theme_bw()

png(paste0("./plots/flowline_regridded.png"), width = 800, height = 800)
print(plot_flowline_rg)
dev.off()
