==> real_data/data_preprocess/01_2_map_elev_data.R <==
## Map surface elevation data to flowline

# library(ncdf4)
# library(CFtime)
library(dplyr)
library(ggplot2)
library(RColorBrewer)
library(lattice)
library(sf) # for shapefiles
# library(sfheaders)
library(sp)
library(parallel)
library(qs)

setwd("~/SSA_model/CNN/real_data/")

data_dir <- "./data"

## Basin shapefile
basin_data <- read_sf(paste0(data_dir, "/boundaries/Basins/Basins_Antarctica_v02.shp"))
thwaites_bound <- basin_data %>% filter(NAME == "Thwaites")

## Grounding line data
gl_thwaites <- qread(paste0(data_dir, "/grounding_line/gl_thwaites.qs"))

# surf_elev_data <- readRDS(file = paste0(data_dir, "/surface_elev/height_thwaites.rds"))
surf_elev_data <- qread(file = paste0(data_dir, "/surface_elev/height_thwaites.qs"))
se_err_grid <- qread(paste0(data_dir, "/surface_elev/height_err_thwaites.qs"))

## Flowline data
delta <- 1920 # grid size
flowline <- qread(paste0(data_dir, "/flowline_regrid.qs"))
flowline_pos <- lapply(1:nrow(flowline), function(i) as.numeric(flowline[i, ]))


## Map surface elevation data to flowline

avg_nearest_four <- function(pos, values) {
    near_pts <- values %>% filter(
        x >= (pos[1] - delta) & x <= (pos[1] + delta),
        y >= (pos[2] - delta) & y <= (pos[2] + delta)
    )

    near_height <- near_pts %>%
        mutate(dist = sqrt((x - pos[1])^2 + (y - pos[2])^2)) %>%
        filter(dist > 0) %>%
        arrange(dist) %>%
        slice_min(dist, n = 4) %>%
        # select(all_of(var)) %>%
        summarise(height_avg = mean(.data[[var]])) # , vx_avg = mean(vx), vy_avg = mean(vy)) # %>% # take average vx and vx as vx and vy at current position
    # as.numeric()

    return(near_height)
}

years <- 2010:2020

se_err_ls <- list()
for (i in 1:length(years)) {
    var <- paste0("err", year)
    # se_err <- se_err_grid %>% select(x, y, all_of(paste0("err", year)))
    se_err_ls[[i]] <- lapply(flowline_pos, avg_nearest_four, values = se_err_grid) # , mc.cores = 12L)
}

qsave(se_err_ls, paste0(data_dir, "/surface_elev/se_err_flowline.qs"))

## Range of error is (0.3170417 1.2102708), so I think it's safe to assume a measurement error sd of about 1m

for (year in years) {
    var <- paste0("height", year)
    cat("Mapping surface elevation data to flowline for", year, "\n")

    annual_surf_elev <- surf_elev_data %>% select(x, y, all_of(var))

    surf_elev_plot <- annual_surf_elev %>% ggplot() + 
                        geom_point(aes(x = x, y = y, color = .data[[var]])) + 
                        scale_colour_distiller(palette = "Blues", direction = 1, 
                                    name = "Height (m)") + 
                        geom_sf(data = thwaites_bound, color = "black", fill = NA) +
                        geom_point(data = gl_thwaites, mapping = aes(x = X, y = Y), color = "salmon", size = 0.05) +
                        geom_point(data = flowline, mapping = aes(x = x, y = y), color = "black", size = 0.1) + 
                        theme_bw()

    png(paste0("./plots/surface_elev/surf_elev_", year, ".png"), width = 800, height = 500)
    print(surf_elev_plot)
    dev.off()

    surf_elev <- lapply(flowline_pos, avg_nearest_four, values = annual_surf_elev) #, mc.cores = 12L)

    png(paste0("./plots/surface_elev/elev_flowline_1d_", year, ".png"), width = 800, height = 500)
    plot(unlist(surf_elev), type = "l")
    dev.off()

    qsave(surf_elev, paste0(data_dir, "/surface_elev/surf_elev_", year, ".qs"))
    qsave(se_err, paste0(data_dir, "/surface_elev/surf_elev_err_", year, ".qs"))
}
